Rendering
This commit is the squash of all the changes I introduced since the
beginning of that experimental project. I had to squash these changes
since they were quite chaotic with lots of backs in forths.
The project in it's current state capable of rendering spheres and
planes with diffuse or metal materials. It can use two type of light
sources: Directional light and Spherical light. Each of them can cast
shadows.